Get software QW for Windows and DOS working.

+++ common/d_surf.c
@@ -80,7 +80,7 @@ D_InitCaches
 
 ================
 */
-void D_InitCaches (void *buffer, int size)
+void D_InitCaches (void *buffer, size_t size)
 {
 	if (!msg_suppress_1)
 		Com_Printf ("%ik surface cache\n", size/1024);
